requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Explainability
grad-cam>=1.4.8

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
